Keep filesystem race tests at semantic boundaries
The containment regression mutates the process working directory and filesystem namespace, so it belongs in the semantic integration test subtree instead of the hermetic unit-test surface. The test behavior and production fix remain unchanged.

Constraint: CI requires effect-bearing filesystem tests to have an explicit semantic integration disposition.
Rejected: Add the test to the migration baseline | that would grow known semantic test debt.
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep this race regression in the semantic integration boundary while it mutates cwd and real filesystem state.
Tested: Focused scaffold suite, 45 steps; semantic disposition audit; formatting; typecheck.
Not-tested: Native Windows symlink race when symlink privilege is unavailable.
